requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.9.1
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-otp==1.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
Markdown==3.8
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pprintpp==0.4.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pyotp==2.9.0
pytest==8.4.1
pytest-clarity==1.0.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-sugar==1.0.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
sqlparse==0.5.3
termcolor==3.1.0
tzdata==2025.2
uritemplate==4.1.1

